Feature: Search program name from environment
The shell searches the process's environment PATH variable for the presence of a program equivalent to the command input(tokens[0]).
The command program found is returned to await process execution.
The $PATH value is tokenized into a linked list and the list is traversed to look for the program name.
Functions related to managing the environment are included in the environment.c file and those of searching the env in the search.c file.
Memory leaks are handled for multiple command inputs however there is need for support of input validation of special characters and delimiters. AAn issue will be opened soon.
Function, struct and library definitons for the functionalities have been added to the shell.h file